+++ src/tests/redirect-uri.test.ts
@@ -1,9 +1,10 @@
 import * as vscode from "vscode";
+import * as validation from "../utils/validation";
 import { GraphApiRepository } from "../repositories/graph-api-repository";
 import { AppRegTreeDataProvider } from "../data/tree-data-provider";
 import { AppRegItem } from "../models/app-reg-item";
 import { RedirectUriService } from "../services/redirect-uri";
-import { mockAppObjectId, seedMockData } from "./data/test-data";
+import { mockApplications, mockAppObjectId, mockRedirectUris, seedMockData } from "./data/test-data";
 import { getTopLevelTreeItem } from "./test-utils";
 
 // Create Jest mocks
@@ -349,6 +350,183 @@ describe("Redirect URI Service Tests", () => {
 		expect(triggerCompleteSpy).toHaveBeenCalled();
 		expect(treeItem?.children?.length).toEqual(2);
 		expect(treeItem?.children?.[1].label).toEqual("https://test.com/callback");
+	});
+
+	test("Add web redirect with successfully with localhost url", async () => {
+		// Arrange
+		const newRedirectUri: string = "http://localhost/callback";
+		item = { objectId: mockAppObjectId, contextValue: "WEB-REDIRECT" };
+		const validationSpy = jest.spyOn(validation, "validateRedirectUri");
+		jest.spyOn(redirectUriService, "inputRedirectUri").mockImplementation(async (item: AppRegItem, existingRedirectUris: string[], validation: (uri: string, context: string, existingRedirectUris: string[], isEditing: boolean, oldValue: string | undefined) => string | undefined) => {
+			const result = validation(newRedirectUri, item.contextValue!, existingRedirectUris, false, undefined);
+			return result === undefined ? newRedirectUri : undefined;
+		});
+
+		// Act
+		await redirectUriService.add(item);
+
+		// Assert
+		const treeItem = await getTopLevelTreeItem(mockAppObjectId, treeDataProvider, "WEB-REDIRECT");
+		expect(statusBarSpy).toHaveBeenCalled();
+		expect(iconSpy).toHaveBeenCalled();
+		expect(validationSpy).toHaveBeenCalled();
+		expect(triggerCompleteSpy).toHaveBeenCalled();
+		expect(treeItem?.children?.length).toEqual(2);
+		expect(treeItem?.children?.[1].label).toEqual(newRedirectUri);
+	});
+
+	test("Edit web redirect with successfully with localhost url", async () => {
+		// Arrange
+		const newRedirectUri: string = "http://localhost/callback";
+		item = { objectId: mockAppObjectId, contextValue: "WEB-REDIRECT", label: mockApplications[0].web.redirectUris[0] };
+		const validationSpy = jest.spyOn(validation, "validateRedirectUri");
+		jest.spyOn(redirectUriService, "inputRedirectUri").mockImplementation(async (item: AppRegItem, existingRedirectUris: string[], validation: (uri: string, context: string, existingRedirectUris: string[], isEditing: boolean, oldValue: string | undefined) => string | undefined) => {
+			const result = validation(newRedirectUri, item.contextValue!, existingRedirectUris, true, item.label!.toString());
+			return result === undefined ? newRedirectUri : undefined;
+		});
+
+		// Act
+		await redirectUriService.edit(item);
+
+		// Assert
+		const treeItem = await getTopLevelTreeItem(mockAppObjectId, treeDataProvider, "WEB-REDIRECT");
+		expect(statusBarSpy).toHaveBeenCalled();
+		expect(iconSpy).toHaveBeenCalled();
+		expect(validationSpy).toHaveBeenCalled();
+		expect(triggerCompleteSpy).toHaveBeenCalled();
+		expect(treeItem?.children?.length).toEqual(1);
+		expect(treeItem?.children?.[0].label).toEqual(newRedirectUri);
+	});
+
+	test("Add spa redirect with successfully with localhost url", async () => {
+		// Arrange
+		const newRedirectUri: string = "http://localhost/callback";
+		item = { objectId: mockAppObjectId, contextValue: "SPA-REDIRECT" };
+		const validationSpy = jest.spyOn(validation, "validateRedirectUri");
+		jest.spyOn(redirectUriService, "inputRedirectUri").mockImplementation(async (item: AppRegItem, existingRedirectUris: string[], validation: (uri: string, context: string, existingRedirectUris: string[], isEditing: boolean, oldValue: string | undefined) => string | undefined) => {
+			const result = validation(newRedirectUri, item.contextValue!, existingRedirectUris, false, undefined);
+			return result === undefined ? newRedirectUri : undefined;
+		});
+
+		// Act
+		await redirectUriService.add(item);
+
+		// Assert
+		const treeItem = await getTopLevelTreeItem(mockAppObjectId, treeDataProvider, "SPA-REDIRECT");
+		expect(statusBarSpy).toHaveBeenCalled();
+		expect(iconSpy).toHaveBeenCalled();
+		expect(validationSpy).toHaveBeenCalled();
+		expect(triggerCompleteSpy).toHaveBeenCalled();
+		expect(treeItem?.children?.length).toEqual(2);
+		expect(treeItem?.children?.[1].label).toEqual(newRedirectUri);
+	});
+
+	test("Add web redirect with duplicate validation error", async () => {
+		// Arrange
+		const newRedirectUri: string = "https://sample.com/callback";
+		item = { objectId: mockAppObjectId, contextValue: "WEB-REDIRECT-URI" };
+		const validationSpy = jest.spyOn(validation, "validateRedirectUri");
+		jest.spyOn(redirectUriService, "inputRedirectUri").mockImplementation(async (item: AppRegItem, existingRedirectUris: string[], validation: (uri: string, context: string, existingRedirectUris: string[], isEditing: boolean, oldValue: string | undefined) => string | undefined) => {
+			const result = validation(newRedirectUri, item.contextValue!, existingRedirectUris, false, undefined);
+			return result === undefined ? newRedirectUri : undefined;
+		});
+
+		// Act
+		await redirectUriService.add(item);
+
+		// Assert
+		expect(redirectUriService.inputRedirectUri).toHaveBeenCalled();
+		expect(validationSpy).toHaveReturnedWith("The Redirect URI specified already exists in this application.");
+		expect(statusBarSpy).toHaveBeenCalled();
+	});
+
+	test("Add web redirect with length validation error", async () => {
+		// Arrange
+		const newRedirectUri: string = "https://sample.com/callback".padEnd(257, "X");
+		item = { objectId: mockAppObjectId, contextValue: "WEB-REDIRECT-URI" };
+		const validationSpy = jest.spyOn(validation, "validateRedirectUri");
+		jest.spyOn(redirectUriService, "inputRedirectUri").mockImplementation(async (item: AppRegItem, existingRedirectUris: string[], validation: (uri: string, context: string, existingRedirectUris: string[], isEditing: boolean, oldValue: string | undefined) => string | undefined) => {
+			const result = validation(newRedirectUri, item.contextValue!, existingRedirectUris, false, undefined);
+			return result === undefined ? newRedirectUri : undefined;
+		});
+
+		// Act
+		await redirectUriService.add(item);
+
+		// Assert
+		expect(redirectUriService.inputRedirectUri).toHaveBeenCalled();
+		expect(validationSpy).toHaveReturnedWith("The Redirect URI is not valid. A Redirect URI cannot be longer than 256 characters.");
+		expect(statusBarSpy).toHaveBeenCalled();
+	});
+
+	test("Add web redirect with incorrect scheme validation error", async () => {
+		// Arrange
+		const newRedirectUri: string = "http://sample.com/callback";
+		item = { objectId: mockAppObjectId, contextValue: "WEB-REDIRECT-URI" };
+		const validationSpy = jest.spyOn(validation, "validateRedirectUri");
+		jest.spyOn(redirectUriService, "inputRedirectUri").mockImplementation(async (item: AppRegItem, existingRedirectUris: string[], validation: (uri: string, context: string, existingRedirectUris: string[], isEditing: boolean, oldValue: string | undefined) => string | undefined) => {
+			const result = validation(newRedirectUri, item.contextValue!, existingRedirectUris, false, undefined);
+			return result === undefined ? newRedirectUri : undefined;
+		});
+
+		// Act
+		await redirectUriService.add(item);
+
+		// Assert
+		expect(redirectUriService.inputRedirectUri).toHaveBeenCalled();
+		expect(validationSpy).toHaveReturnedWith("The Redirect URI is not valid. A Redirect URI must start with https:// unless it is using http://localhost.");
+		expect(statusBarSpy).toHaveBeenCalled();
+	});
+
+	test("Add native redirect with incorrect scheme validation error", async () => {
+		// Arrange
+		const newRedirectUri: string = "not/a/url";
+		item = { objectId: mockAppObjectId, contextValue: "NATIVE-REDIRECT" };
+		const validationSpy = jest.spyOn(validation, "validateRedirectUri");
+		jest.spyOn(redirectUriService, "inputRedirectUri").mockImplementation(async (item: AppRegItem, existingRedirectUris: string[], validation: (uri: string, context: string, existingRedirectUris: string[], isEditing: boolean, oldValue: string | undefined) => string | undefined) => {
+			const result = validation(newRedirectUri, item.contextValue!, existingRedirectUris, false, undefined);
+			return result === undefined ? newRedirectUri : undefined;
+		});
+
+		// Act
+		await redirectUriService.add(item);
+
+		// Assert
+		expect(redirectUriService.inputRedirectUri).toHaveBeenCalled();
+		expect(validationSpy).toHaveReturnedWith("The Redirect URI is not valid. A Redirect URI must start with https, http, or customScheme://.");
+		expect(statusBarSpy).toHaveBeenCalled();
+	});
+
+	test("Add web redirect for AAD audiences with total error", async () => {
+		// Arrange
+		item = { objectId: mockAppObjectId, contextValue: "WEB-REDIRECT" };
+		const warningSpy = jest.spyOn(vscode.window, "showWarningMessage");
+		jest.spyOn(graphApiRepository, "getApplicationDetailsPartial").mockImplementation(async (_id: string, _select: string, _expandOwners?: boolean | undefined) => ({ success: true, value: mockRedirectUris }));
+		jest.spyOn(graphApiRepository, "getSignInAudience").mockImplementation(async (_id: string) => ({ success: true, value: "AzureADMyOrg" }));
+
+		// Act
+		await redirectUriService.add(item);
+
+		// Assert
+		expect(statusBarSpy).toHaveBeenCalled();
+		expect(iconSpy).toHaveBeenCalled();
+		expect(warningSpy).toHaveBeenCalledWith("You cannot add any more Redirect URIs. The maximum for this application type is 256.", "OK");
+	});
+
+	test("Add web redirect for AAD and consumer audiences with total error", async () => {
+		// Arrange
+		item = { objectId: mockAppObjectId, contextValue: "WEB-REDIRECT" };
+		const warningSpy = jest.spyOn(vscode.window, "showWarningMessage");
+		jest.spyOn(graphApiRepository, "getApplicationDetailsPartial").mockImplementation(async (_id: string, _select: string, _expandOwners?: boolean | undefined) => ({ success: true, value: mockRedirectUris }));
+		jest.spyOn(graphApiRepository, "getSignInAudience").mockImplementation(async (_id: string) => ({ success: true, value: "AzureADandPersonalMicrosoftAccount" }));
+
+		// Act
+		await redirectUriService.add(item);
+
+		// Assert
+		expect(statusBarSpy).toHaveBeenCalled();
+		expect(iconSpy).toHaveBeenCalled();
+		expect(warningSpy).toHaveBeenCalledWith("You cannot add any more Redirect URIs. The maximum for this application type is 100.", "OK");
 	});
 
 	test("Add web redirect successfully for AAD and consumer audiences", async () => {